requirements.txt
pandas>=1.5
numpy
scikit-learn>=1.0
pycaret[full]>=3.0.0
streamlit>=1.20
matplotlib
ydata-profiling
transformers>=4.30.0
torch
sentencepiece
python-dotenv
joblib
